refactor: 결제 코드 리팩토링
