+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/subscription/controller/ViewSubscriptionController.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 /**
- * 요금제를 프론트에 표시하기 위한 화면 전환용 Controller
+ * 요금제 화면용 컨트롤러
  *
  * @author : 이지헌
  * @filename : ViewSubscriptionController
