+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/subscription/repository/SubscriptionRepository.java
@@ -4,7 +4,7 @@ import com.multi.mlpenterpriseapprovalsystem.subscription.domain.Subscription;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
- * 요금제 테이블 접근 repository
+ * 요금제 테이블 관리 레포지토리
  *
  * @author : 이지헌
  * @filename : SubscriptionRepository
